Alchemist Prototyping (Crafting Items)
Created scripts that detect when an object is picked up, when it is an INT value goes up and the object is destroyed. The INT value will be what is placed inside the atendee menu so players can visually see how many objects they have collected.

System works up until crafting item 1